cqrs_embedding_gen_svc/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
sentence-transformers>=2.3.0
huggingface_hub>=0.16.0
PyYAML==6.0.1
numpy==1.24.3
requests==2.31.0
transformers>=4.32.0
torch>=2.0.0